backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai==0.4.0